+++ v3_00/main.py
@@ -61,6 +61,11 @@ class Game:
         # -- scene elements --  
         self.scene_pinboard_image = pg.image.load(path.join(imgs_folder, SCENE_INFO_PINBOARD_IMG)).convert_alpha()  
         self.scene_pinboard_paper_image = pg.image.load(path.join(imgs_folder, SCENE_PINBOARD_PAPER_IMG)).copy().convert_alpha() 
+        self.emoji_1_img = pg.image.load(path.join(imgs_folder, SCENE_PINBOARD_ICON_1_IMG)).copy().convert_alpha() 
+        self.emoji_2_img = pg.image.load(path.join(imgs_folder, SCENE_PINBOARD_ICON_2_IMG)).copy().convert_alpha() 
+        self.emoji_3_img = pg.image.load(path.join(imgs_folder, SCENE_PINBOARD_ICON_3_IMG)).copy().convert_alpha() 
+        self.emoji_4_img = pg.image.load(path.join(imgs_folder, SCENE_PINBOARD_ICON_4_IMG)).copy().convert_alpha() 
+        self.emoji_5_img = pg.image.load(path.join(imgs_folder, SCENE_PINBOARD_ICON_5_IMG)).copy().convert_alpha() 
         # - windows -      
         self.window_img = pg.image.load(path.join(imgs_folder, WINDOW_IMG)).convert_alpha()  
         self.window_border_img = pg.image.load(path.join(imgs_folder, WINDOW_BORDER_1_IMG)).convert_alpha()  
@@ -70,6 +75,9 @@ class Game:
         self.window_hl_2_img = pg.image.load(path.join(imgs_folder, WINDOW_HL_2_IMG)).convert_alpha()  
         self.window_shelved_1_img = pg.image.load(path.join(imgs_folder, WINDOW_SHELVED_1_IMG)).convert_alpha()  
         self.window_shelved_hl_1_img = pg.image.load(path.join(imgs_folder, WINDOW_SHELVED_HL_1_IMG)).convert_alpha()  
+        # - tab bar elements -
+        self.tab_bar_prep_img = pg.image.load(path.join(imgs_folder, TAB_BAR_PREPARING_IMG)).convert_alpha()  
+        self.tab_bar_order_img = pg.image.load(path.join(imgs_folder, TAB_BAR_ORDERING_IMG)).convert_alpha()  
         # - chat elements -
         self.payment_pending_1_img = pg.image.load(path.join(imgs_folder, PAYMENT_PENDING_IMG_1)).convert_alpha()  
         self.payment_success_1_img = pg.image.load(path.join(imgs_folder, PAYMENT_SUCCESS_IMG_1)).convert_alpha()  
@@ -88,11 +96,20 @@ class Game:
         self.FONT_BOHEMIAN_TYPEWRITER_26 = pg.font.Font((path.join(fonts_folder, "Bohemian Typewriter.ttf")), 26)
         self.FONT_BOHEMIAN_TYPEWRITER_32 = pg.font.Font((path.join(fonts_folder, "Bohemian Typewriter.ttf")), 32)
         self.FONT_BOHEMIAN_TYPEWRITER_46 = pg.font.Font((path.join(fonts_folder, "Bohemian Typewriter.ttf")), 46)
+        # -- load more fonts -- 
+        self.FONT_LATO_16 = pg.font.Font((path.join(fonts_folder, "Lato-Black.ttf")), 16) 
+        self.FONT_LATO_20 = pg.font.Font((path.join(fonts_folder, "Lato-Black.ttf")), 20) 
         # -- define main gui surface dimensions --
-        self.pc_screen_surf_width, self.pc_screen_surf_height = 1000, 600
-        self.pc_screen_surf_x, self.pc_screen_surf_y = (WIDTH / 2) - (self.pc_screen_surf_width / 2), 100
+        self.pc_screen_surf_width, self.pc_screen_surf_height = 1000, 620
+        self.pc_screen_surf_x, self.pc_screen_surf_y = (WIDTH / 2) - (self.pc_screen_surf_width / 2), 100 # testing +15 for tab bar adjustment
         self.tab_bar_height = 50 # the top bar on the pc_screen_surf the emulate browser tabs
         self.pc_screen_surf_true_y = self.pc_screen_surf_y + self.tab_bar_height # else y val doesnt take the tab_bar_height into consideration
+        
+        # [new!] - preloading core scene gui vars for start screen
+        self.waiting_start = True
+        self.pinboard_image_surf = self.scene_pinboard_image.copy()
+        self.login_1_img = pg.image.load(path.join(imgs_folder, START_LOGIN_IMG_1)).convert_alpha()  
+
 
     def new_level(self):
         """ initialize all variables and do all the setup for a new game """
@@ -126,6 +143,38 @@ class Game:
         self.pinboard_pos = (0, 20)   
         self.customer_sidebar_queue = {}   
        
+
+    # [ new! ]
+    # -- test - for run start --
+    # - guna move this stuff and group it together -
+    
+    def run_start(self):
+        while self.waiting_start:
+            self.dt = self.clock.tick(FPS) / 1000.0
+            want_start = self.render_start_screen()
+            if want_start:
+                self.waiting_start = False
+
+    def draw_start(self):
+        pg.display.set_caption(f"Crud Cafe v3.11 - {self.clock.get_fps():.2f}")
+        # -- draw the background and screen surf -- 
+        self.scene_img.blit(self.pinboard_image_surf, (0,20))
+        self.screen.blit(self.scene_img, (0, 0)) 
+        self.wipe_computer_screen_surface()
+        self.pc_screen_surf.blit(self.login_1_img, (0, 0)) # 50 is the top tabs area, need to hard code this once added it in 
+        # -- redraw the screen once we've blit to it, with a rect as a temp faux monitor outline/edge --
+        screen_outline_rect = self.screen.blit(self.pc_screen_surf, (self.pc_screen_surf_x, self.pc_screen_surf_y))
+        screen_outline_rect.height += 15
+        pg.draw.rect(self.screen, DARKGREY, screen_outline_rect, 25) # draws the faux monitor edge around the screen surf     
+        # --
+        
+
+        # --
+        pg.display.flip()
+
+    # -- end new test for start screen stuff, which will move shortly --
+
+
     def run(self):
         # runs the game loop... thank you for coming to my TEDtalk
         self.playing = True
@@ -141,7 +190,11 @@ class Game:
     
     def render_start_screen(self):
         """ render the start screen... thank you for coming to my TEDtalk """
-        ...
+        self.draw_start()
+        want_start = self.wait_for_continue()
+        if want_start:
+            return True
+
 
     def game_over_man_game_over(self):
         """ render the game over screen """
@@ -153,9 +206,7 @@ class Game:
     def draw(self):
         pg.display.set_caption(f"Crud Cafe v3.11 - {self.clock.get_fps():.2f}")
         # -- draw the background -- 
-        self.screen.blit(self.scene_img, (0,0)) 
-        # -- [new!] --
-        self.pinboard_image_surf = self.scene_pinboard_image.copy() # - added this to stop the over blit issues -
+        self.screen.blit(self.scene_img, (0, 0)) 
         # -- [new!] - draw to the pinboard image, just basic setup stuff dw too much about the vars n states yet --
         self.write_info_counters_to_pinboard(font_size=46, post_it="ordering")
         self.write_info_counters_to_pinboard(font_size=46, post_it="cancelled")
@@ -179,6 +230,11 @@ class Game:
         self.scene_img.blit(self.pinboard_image_surf, self.pinboard_pos)
         # -- wipe the computer screen surface at the start of each frame, we then draw to this surface and then blit it to the screen (without the fill) -- 
         self.wipe_computer_screen_surface()
+        
+        # [ new! ] 
+        # -- draw tab bar img --
+        self.draw_tab_bar()
+
         # -- loop tabs --
         for sprite in self.browser_tabs:
             if isinstance(sprite, Browser_Tab): # really for type hinting
@@ -202,12 +258,17 @@ class Game:
                     # -- finally, run this for all child instances, if they are they active tab you draw their surface to the scene screen image --                     
                     sprite.draw_tab_to_pc()       
 
+        # # [ new! ] 
+        # # -- draw tab bar img --
+        # self.draw_tab_bar()
+
         # [ new! ] 
         # -- new first implementation of preparing orders tab & its functionality --
         self.preparing_orders_tab.draw()          
 
         # -- redraw the screen once we've blit to it, with a rect as a temp faux monitor outline/edge --
         screen_outline_rect = self.screen.blit(self.pc_screen_surf, (self.pc_screen_surf_x, self.pc_screen_surf_y))
+        screen_outline_rect.height += 15
         pg.draw.rect(self.screen, DARKGREY, screen_outline_rect, 25) # draws the faux monitor edge around the screen surf               
         # -- finally, flip the display --1
         pg.display.flip()
@@ -295,6 +356,12 @@ class Game:
         """ we use this as a base to draw everything on to, it is basically our game surface """
         self.pc_screen_surf = pg.Surface((self.pc_screen_surf_width, self.pc_screen_surf_height))
         self.pc_screen_surf.fill(SKYBLUE) # once setup the actual tab bar, make this black - then will make boot up / boot down animation hella easy
+
+    def draw_tab_bar(self):
+        if self.new_orders_tab.is_active_tab:
+            self.pc_screen_surf.blit(self.tab_bar_order_img.copy(), (25, 25))
+        else:
+            self.pc_screen_surf.blit(self.tab_bar_prep_img.copy(), (25, 25))
 
 
     # ---- For Pinboard Scene GUI ----
@@ -398,12 +465,22 @@ class Game:
             if self.preparing_orders_tab.tab_scroll_offset > 0:
                 self.preparing_orders_tab.tab_scroll_offset = 0 # coube achieved with min() too tbf
 
+    # -- [ new! ] - for game start / game over --
+    def wait_for_continue(self):
+        pg.event.wait()
+        wait_input = True
+        while wait_input:
+            self.clock.tick(FPS)
+            for event in pg.event.get():
+                if event.type == pg.KEYUP:
+                    return "start"
+
 # ---- End Game Class ----
 
 
 # -- instantiate a new game object and run the game --
 g = Game()
-g.render_start_screen()
+g.run_start()
 while True:
     g.new_level()
     g.run()